backend/requirements.txt
fastapi
uvicorn
google-generativeai
chromadb
websockets
python-dotenv
python-multipart
Pillow
